pyproject.toml
[tool.poetry]
name = "django-authorization"
version = "0.1.0"
description = ""
authors = ["Suvan <overmind77s@gmail.com>"]
readme = "README.md"
packages = [{include = "django_authorization"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-configurations = "^2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
